refactor: getScripts
