--- packages/web3-providers/src/Twitter/apis/getHeaders.ts
@@ -1,15 +1,0 @@
-import { getTokens } from './getTokens.js'
-
-export async function getHeaders(overrides?: Record<string, string>) {
-    const { bearerToken, csrfToken } = await getTokens()
-
-    return {
-        authorization: `Bearer ${bearerToken}`,
-        'x-csrf-token': csrfToken,
-        'content-type': 'application/json',
-        'x-twitter-auth-type': 'OAuth2Session',
-        'x-twitter-active-user': 'yes',
-        referer: 'https://twitter.com/',
-        ...overrides,
-    }
-}
