+++ packages/web3-providers/src/Twitter/apis/index.ts
@@ -1,7 +1,6 @@
 export * from './getDatabase.js'
 export * from './getSettings.js'
 export * from './getTokens.js'
-export * from './getHeaders.js'
 export * from './getUserNFTContainer.js'
 export * from './getUserSettings.js'
 export * from './getUserViaTwitterIdentity.js'
